Fix SelectedElementType access by keeping z-order helper local
